# SANGAM-PRODUCTION/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SANGAM — Supply Chain Management API",
    "version": "1.0.0",
    "description": "REST API for the SANGAM permissioned blockchain-based supply chain\nmanagement system built for the Indian Army.\n\n## Authentication\nAll protected endpoints require a `Bearer` JWT obtained from `POST /api/auth/login`.\nTokens expire after 8 hours (one operational shift). Use `POST /api/auth/refresh` to rotate.\n\n## Permissions\nEvery endpoint lists the required permission (format: `resource:action`).\nPermissions are granted by role (Day 13 RBAC), temporary delegation (Day 15),\nor single-use emergency override (Day 15).\n\n## Command Scope\nData is filtered to the requesting user's command scope — their unit and all\nsubordinate units. A battalion CO automatically sees all companies and platoons.\nLateral access (sibling units) is denied.\n\n## Offline-First\nAll services degrade gracefully when the database is unavailable.\nNotifications and audit events are buffered in memory and flushed on reconnect.",
    "contact": {
      "name": "SANGAM Development Team",
      "url": "https://github.com/sangam-scm"
    },
    "license": {
      "name": "Proprietary — Indian Army Internal Use"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local / Docker development"
    },
    {
      "url": "https://sangam.army.mil",
      "description": "Production (classified network)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from POST /api/auth/login (Day 14). Expires after 8h."
      }
    },
    "schemas": {
      "Success": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "INVALID_REQUEST"
          },
          "message": {
            "type": "string",
            "example": "A descriptive error message"
          }
        }
      },
...
      "NotificationType": {
        "type": "string",
        "enum": [
          "LOW_STOCK",
          "TRANSFER_PENDING",
          "TRANSFER_APPROVED",
          "TRANSFER_REJECTED",
          "MESH_PEER_OFFLINE",
          "MESH_PEER_ONLINE",
          "SYNC_CONFLICT",
          "SECURITY_ALERT",
          "BLOCKCHAIN_TAMPER",
          "SYSTEM_ANNOUNCEMENT",
          "DELEGATION_GRANTED"
        ]
      },
      "Severity": {
        "type": "string",
        "enum": [
          "LOW",
          "MEDIUM",
          "HIGH",
          "CRITICAL"
        ]
      },
      "Notification": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "type": {
            "$ref": "#/components/schemas/NotificationType"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "title": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "sourceUnitId": {
            "type": "integer",
            "nullable": true
          },
          "minRankLevel": {
            "type": "integer"
          },
          "targetUserId": {
            "type": "integer",
            "nullable": true
          },
          "resourceType": {
            "type": "string",
            "nullable": true
          },
          "resourceId": {
            "type": [
              "integer",
              "string"
            ],
            "nullable": true
          },
          "requiresAck": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "read": {
            "type": "boolean",
            "description": "Per-user read status"
          },
          "acknowledged": {
            "type": "boolean",
            "description": "Per-user acknowledge status"
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid JWT",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions or outside command scope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/notifications": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "List visible notifications",
        "description": "Returns notifications visible to the current user based on rank and command scope. Filters available: unreadOnly, type, severity.",
        "operationId": "listNotifications",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "unreadOnly",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/NotificationType"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Severity"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notification list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "notifications": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Notification"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "unreadTotal": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/notifications/unread-count": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Unread notification count",
        "description": "Returns the unread count for the notification badge.",
        "operationId": "getUnreadCount",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Unread count",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "unreadCount": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/notifications/digest": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Daily digest",
        "description": "Summary of notifications in a time window — counts by type/severity and pending acknowledgments.",
        "operationId": "getDigest",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "hours",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 24
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Digest",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/notifications/stream": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Real-time notification stream (SSE)",
        "description": "Server-Sent Events stream delivering notifications in real time. Connect with EventSource and include a JWT token as query param `?token=...`.",
        "operationId": "streamNotifications",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "text/event-stream — notifications as JSON data events",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/notifications/{id}/acknowledge": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Acknowledge a notification",
        "description": "Mark a notification as acknowledged (implies read). For requiresAck notifications, this is an auditable action.",
        "operationId": "acknowledgeNotification",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Acknowledged",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "security": []
}